Add files via upload
The aim of this project is to use logistic regression to predict whether a tumor is benign or
malignant based on diagnostic features of breast cancer tumors using the Breast Cancer Wisconsin
(Diagnostic) dataset. The data is first preprocessed by cleaning, scaling, and encoding the features
as necessary for logistic regression. The performance of the model is evaluated using various
metrics including accuracy, ROC AUC score, precision, recall, and F1-score. Additionally, feature
selection and interpretation of the logistic regression model are performed to gain insights into the
relationship between the input features and the target variable. Hyperparameter tuning is carried
out to improve the performance of the model. The results show that the logistic regression model
achieves a high level of accuracy and ROC AUC score in predicting whether a tumor is benign or
malignant. The feature selection and interpretation also provide valuable insights into the important
features that contribute to the classification task. Moreover, the effectiveness of logistic regression
is compared with other models such as SVM, random forest, and decision tree. Regularization
techniques such as L1 and L2 regularization are also implemented. Our analysis resulted in an
accuracy of 0.98 for the logistic regression model after PCA. Overall, this project highlights the
importance of using logistic regression for classification tasks in healthcare and the significance of
feature selection and interpretation in gaining insights into the underlying patterns in the data.